+++ package.json
@@ -17,6 +17,7 @@
     "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
     "react-h5-audio-player": "^3.8.6",
+    "react-icons": "^4.7.1",
     "react-id-generator": "^3.0.2",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.6.1",
@@ -24,6 +25,7 @@
     "react-uuid": "^2.0.0",
     "styled-components": "^5.3.6",
     "typescript": "^4.9.4",
+    "wavesurfer.js": "^6.4.0",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
